IMDB-Scrapper is a tiny tool that grabs all links from IMDB Top 250 movies and then scrapes their data.
IMDB-Scrapper is a tiny tool that grabs all links from IMDB Top 250 movies and then scrapes their data. The tool uses GET requests because we don't have the imdb API in hand.

	Which datas the tool scrapes ?
		- Movie title
		- Movie rating
		- Year
		- Genre
		- Director
		- Movie stars
		- Summary of movie
	Used modules
		-requests
		-BeatifulSoup
		-Pandas
	Saving formats
		-txt
		-csv
	Usage
		- Just run imdb_scraper.py :)

This repository includes datas which i scraped with this tool are in the main folder and free to use for all data-science practitioner and students. I decided to upload it maybe someone needs a sample data for practicing.

The tool is core of my future fun projects and implementations.